Move attribute descriptions into attribute service
