+++ SWLOR.Game.Server.Tests/Feature/CombatUpgradeTlkConsistencyTests.cs
@@ -2,6 +2,7 @@ using System.Text.Json;
 using FluentAssertions;
 using NUnit.Framework;
 using SWLOR.Game.Server.Service;
+using SWLOR.Game.Server.Service.AttributeService;
 
 namespace SWLOR.Game.Server.Tests.Feature;
