+++ SWLOR.Game.Server/Feature/GuiDefinition/CharacterStatRebuildDefinition.cs
@@ -1,5 +1,6 @@
 using SWLOR.Game.Server.Feature.GuiDefinition.ViewModel;
 using SWLOR.Game.Server.Service;
+using SWLOR.Game.Server.Service.AttributeService;
 using SWLOR.Game.Server.Service.GuiService;
 
 namespace SWLOR.Game.Server.Feature.GuiDefinition
